fix: remove unnecesarry cards
