+++ src/components/ProjectCollaborators/types.ts
@@ -13,4 +13,14 @@ export interface CardProps {
   project?: ProjectType;
 }
 
+export interface ProjectMetaDataDialogProps
+  extends React.HTMLAttributes<HTMLDivElement> {
+  project?: any;
+  selectedUser?: any;
+  request?: string;
+  roleId?: string;
+  open?: boolean;
+  setOpen?: (value: boolean) => void;
+}
+
 export interface ProjectInvitationProps {}
